requirements.txt
requests
schedule
pymongo
bs4
retrying
python-dotenv
imgurpython
aiogram